+++ sdk/web-pubsub/web-pubsub/samples-dev/managingGroups.ts
@@ -16,7 +16,6 @@ const adminGroup = chatHub.group("admin");
 async function main() {
   // adding and removing users
   await adminGroup.addUser("bterlson");
-  await adminGroup.hasUser("bterlson"); // true
   await adminGroup.removeUser("xirzec");
 
   // adding and removing specific connections